requirements.txt
babel==2.16.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
feedparser==6.0.11
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.5
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
mkdocs-git-revision-date-localized-plugin==1.2.6
mkdocs-material==9.5.31
mkdocs-material-extensions==1.3.1
mkdocs-redirects==1.2.1
mkdocs-section-index==0.3.9
packaging==24.1
paginate==0.5.6
pathspec==0.12.1
platformdirs==4.2.2
Pygments==2.18.0
pymdown-extensions==10.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.7.24
requests==2.32.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
urllib3==2.2.2
watchdog==4.0.2
